Docker environment: build static ipmitool
ipmitool is used for managing IPMI (server management) cards, and can
also manage some functions of Dell iDRAC cards.  VMWare ESXi does not
come with ipmitool, but is able to run a Linux ipmitool if it is
statically linked.

This environment uses a CentOS 4.9 32-bit (i386) base to create a build
environment for a statically linked ipmitool.static that can be copied
onto VMWare ESXi hosts.  (Tested with VMWare ESXi 5.0)